+++ backend/src/main/java/com/wefine/app/service/ArticleService.java
@@ -9,11 +9,11 @@ public interface ArticleService {
 
     List<Article> getAllArticles();
 
-    Article getArticleById(int articleId);
+    Article getArticleById(Long articleId);
 
     boolean createArticle(Article article);
 
     void updateArticle(Article article);
 
-    void deleteArticle(int articleId);
+    void deleteArticle(Long articleId);
 }
